+++ src/test/java/com/alipay/oceanbase/hbase/secondary/OHTableSecondaryPartCheckAndMutateTest.java
@@ -1,0 +1,93 @@
+/*-
+ * #%L
+ * OBKV HBase Client Framework
+ * %%
+ * Copyright (C) 2025 OceanBase Group
+ * %%
+ * OBKV HBase Client Framework  is licensed under Mulan PSL v2.
+ * You can use this software according to the terms and conditions of the Mulan PSL v2.
+ * You may obtain a copy of Mulan PSL v2 at:
+ *          http://license.coscl.org.cn/MulanPSL2
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+ * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+ * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+ * See the Mulan PSL v2 for more details.
+ * #L%
+ */
+
+package com.alipay.oceanbase.hbase.secondary;
+
+import com.alipay.oceanbase.hbase.OHTableClient;
+import com.alipay.oceanbase.hbase.util.ObHTableTestUtil;
+import com.alipay.oceanbase.hbase.util.TableTemplateManager;
+import org.apache.hadoop.hbase.client.*;
+import org.apache.hadoop.hbase.filter.CompareFilter;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+
+import static com.alipay.oceanbase.hbase.util.ObHTableSecondaryPartUtil.*;
+import static com.alipay.oceanbase.hbase.util.ObHTableTestUtil.FOR_EACH;
+import static org.apache.hadoop.hbase.util.Bytes.toBytes;
+import static org.junit.Assert.assertEquals;
+
+
+public class OHTableSecondaryPartCheckAndMutateTest {
+    private static List<String> tableNames = new LinkedList<String>();
+    private static Map<String, List<String>> group2tableNames = new LinkedHashMap<>();
+
+
+    @BeforeClass
+    public static void before() throws Exception {
+        for (TableTemplateManager.TableType type : TableTemplateManager.TableType.values()) {
+            createTables(type, tableNames, group2tableNames, true);
+        }
+    }
+
+    @AfterClass
+    public static void finish() throws Exception {
+        dropTables(tableNames, group2tableNames);
+    }
+
+    @Before
+    public void prepareCase() throws Exception {
+        truncateTables(tableNames, group2tableNames);
+    }
+    
+    public static void testCheckAndMutateImpl(String tableName) throws Exception {
+        OHTableClient hTable = ObHTableTestUtil.newOHTableClient(getTableName(tableName));
+        hTable.init();
+
+        String family = getColumnFamilyName(tableName);
+        String key = "putKey";
+        String column = "putColumn";
+        String value = "value";
+        String newValue = "newValue";
+        RowMutations mutations = new RowMutations(key.getBytes());
+        Put put = new Put(key.getBytes());
+        long timestamp = System.currentTimeMillis();
+        put.add(family.getBytes(), column.getBytes(), timestamp, toBytes(value));
+        hTable.put(put);
+
+        Put newPut = new Put(key.getBytes());
+        newPut.add(family.getBytes(), column.getBytes(), timestamp, toBytes(newValue));
+        mutations.add(newPut);
+        hTable.checkAndMutate(key.getBytes(), family.getBytes(),
+                column.getBytes(), CompareFilter.CompareOp.EQUAL, value.getBytes(),
+                mutations);
+
+        hTable.close();
+    }
+    
+    @Test
+    public void testCheckAndMutate() throws Exception {
+        FOR_EACH(tableNames, OHTableSecondaryPartCheckAndMutateTest::testCheckAndMutateImpl);
+    }
+
+}
